web_app_edition/backend/requirements.txt

channels==2.2.0  
channels-redis==2.4.0        
Django==2.2.4     
django-cors-headers==3.1.0     
djangorestframework==3.10.2    
djangorestframework-jwt==1.11.0      
orderedset==2.0.3     
redis==3.3.8     
psycopg2-binary==2.8.4 

